[REF] hr_holidays: split files according to guidelines
Splits `resource.py` into
- `resource_calendar_leaves.py`
- `resource_calendar.py`
- `resource_resource.py`

No related PR
