+++ addons/hr_holidays/models/resource_calendar.py
@@ -1,0 +1,20 @@
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import fields, models
+
+
+class ResourceCalendar(models.Model):
+    _inherit = "resource.calendar"
+
+    associated_leaves_count = fields.Integer("Time Off Count", compute='_compute_associated_leaves_count')
+
+    def _compute_associated_leaves_count(self):
+        leaves_read_group = self.env['resource.calendar.leaves']._read_group(
+            [('resource_id', '=', False), '|', ('calendar_id', 'in', self.ids), ('calendar_id', '=', False)],
+            ['calendar_id'],
+            ['__count'],
+        )
+        result = {calendar.id if calendar else 'global': count for calendar, count in leaves_read_group}
+        global_leave_count = result.get('global', 0)
+        for calendar in self:
+            calendar.associated_leaves_count = result.get(calendar.id, 0) + global_leave_count
